feat: governance utils
