+++ typescript/nomad-deploy/src/index.ts
@@ -2,6 +2,7 @@ export * as core from './core';
 export * as bridge from './bridge';
 export * as incremental from './incremental';
 
+export * as governance from './governance';
 export * as deploy from './deploy';
 export * as chain from './chain';
 export * as contracts from './contracts';
