+++ typescript/nomad-deploy/tsconfig.json
@@ -16,5 +16,6 @@
     "./src/bridge/*.ts",
     "./src/verification/*.ts",
     "./src/incremental/*.ts",
+    "./src/governance/**/*.ts"
   ]
 }
